+++ src/components/Main/Softwares/Software/Details/Details.data.js
@@ -1,0 +1,8 @@
+export const urls = {
+  testCaseFileManager: {
+    video: "https://youtu.be/pIgJa1uZ48c",
+    exe: "https://athena-web-bucket.s3.eu-west-1.amazonaws.com/test_case_file_manager/test_case_file_manager_v106_setup.exe",
+    doc: "https://athena-web-bucket.s3.eu-west-1.amazonaws.com/test_case_file_manager/tcfm_user_documentation.pdf",
+  },
+  carbonTracker: { video: "", exe: "", doc: "" },
+};